Create Post
Create a file in to the admin/posts/ dir to create.blade.php file. Create a admin-sidebar-posts-links file to the components dir.
Changes to the .env file to the local storage path to the public folder using php artisan storage:link command. And also same change the .env.example file